PythonProject/requirements.txt
graphics.py
numpy